refactor(iam): extract evaluator inline tests to sibling
evaluator.rs grew to 2895 lines, mostly inline #[cfg(test)] mod tests {}
(1019-2895, 1874 lines). Move test body to evaluator_tests.rs sibling
and reference via #[path].

evaluator.rs: 2895 -> 1019 lines (-65%). 378 lib tests still pass.
